Fix broken links in Arcade Build Doco
The arcade cabinet build instructions direct users to a broken link on the Splashkit.io docs site.
The URL points to "https://splashkit.io/articles/installation/ubuntu/"

But this page as changed to https://splashkit.io/installation/linux/ at some point. It's possible to navigate to the correct page from the 404 in a couple of clicks, but there's no reason not to update this URL to point to the correct place.

Also replaced an ugly full link to sharepoint with an embedded link for better readability.

There's also a broken image in the "Configure Start CLI on Boot and Network Manager" section that pointed to a typo'd file that I have corrected.

The other links seem to be free from link rot.